RES-288 - Introduce Database

See merge request futuretechnologies/tornet-jonstaka/trading-platform-poc!136

+++ main.py
@@ -1,14 +1,17 @@
+import argparse
 import os
 import pickle
 from logging.handlers import TimedRotatingFileHandler
 
 from pkg_resources import resource_filename
-from tradingplatformpoc.app.app_inputs import read_config
 import logging
 import sys
+from tradingplatformpoc.connection import SessionMaker
 from tradingplatformpoc.constants import MOCK_DATA_PATH
+from tradingplatformpoc.database import create_db_and_tables, insert_default_config_into_db
 
 from tradingplatformpoc.simulation_runner.trading_simulator import TradingSimulator
+from tradingplatformpoc.sql.job.crud import delete_job, get_job_id_for_config
 
 # --- Read sys.argv to get logging level, if it is specified ---
 string_to_log_later = None
@@ -48,12 +51,25 @@ if string_to_log_later is not None:
 # --- Define path to mock data
 mock_datas_path = resource_filename("tradingplatformpoc.data", "mock_datas.pickle")
 results_path = "./results/"
-config_data = read_config(name='default')
+parser = argparse.ArgumentParser()
+parser.add_argument("--config_id", dest="config_id", default="default",
+                    help="Config ID", type=str)
+args = parser.parse_args()
+
+# config_data = read_config(name=args.config_name)
 
 if __name__ == '__main__':
-    logger.info("Running main")
-    simulator = TradingSimulator(read_config(), MOCK_DATA_PATH)
-    simulation_results = simulator.run()
-    logger.info("Finished running simulations. Will save simulations results as a pickle file.")
-    with open(results_path + 'simulation_results.pickle', 'wb') as destination:
-        pickle.dump(simulation_results, destination)
+    logger.info("Running main with config {}.".format(args.config_id))
+    create_db_and_tables()
+    insert_default_config_into_db()
+    with SessionMaker() as sess:
+        job_id = get_job_id_for_config('default', sess)
+    delete_job(job_id)
+    simulator = TradingSimulator(args.config_id, MOCK_DATA_PATH)
+    simulation_results = simulator()
+    if simulation_results is not None:
+        logger.info("Finished running simulations. Will save simulations results as a pickle file.")
+        with open(results_path + 'simulation_results.pickle', 'wb') as destination:
+            pickle.dump(simulation_results, destination)
+    else:
+        logger.info("Could not finish simulation.")
